Implement basic RL env
Implemented a basic RL env similar to the paper Privacy-Cost Manaagement in Smart Meters With Mutual-Information-Based Reinforcement learning, with a pre-trained H-network as a placeholder (instead of training the H-network along with the env)

Passed the env check provided by gym.

The DDQL-network is left to be implemented.

Also, the function of selecting different episodes as initial states are left to be implemented.

If possible, we can implement a visualizer for per-step visualization. We can draw a line graph to show the user's load and grid load as the env is running. This can be very useful in debug or showcasing our work.