Implement PPO agent and basic rendering
Implement a PPO agent to train it with a pre-trained H-network

Also implemented a baisc rendering function which can show the grid load and user's load, also the change of state-of-charge of the rechargeable battery.

Git commit before implementing an advanced version of rendering which will create a window using a different process, then send the created graph to that process and then show it.