Edit reward func to closer implementation of the paper
Amend the reward func and names of corresponding functions (in the env) to reflect the paper's original design. Now the reward function should behave alike the paper in theory. In another words, the privacy signal reflects the Mutual Information calculation.

However, the privacy signal is too large that it out-weights the cost signal. Further improvement is required on the ration between two signals, and the reduction mode of the criterion (loss func).

At least, in theory, the paper is correct, and it can be roughly implemented. However, our experiment setup has a higher sample rate (1/6 Hz) than the paper (1/(15 * 60) Hz). The total number of samples in an episode is much higher such that the RNN (LSTM) cannot take all samples in one go. The window-size is too small.

This deviates the setup in the paper, in which the LSTM is fed with full-episode data (i.e. 96 samples of a day in the env), which makes the loss returned approximates MI.

In our current implementation, with a LSTM with sequence-length = 512 samples, it only computes MI of the past 51.2 minutes. There is more work to do to fully apply the insight of the paper.