Amend f-signal calculation
f-signal now considers the whole sequence, as we break the previous timestep data into chunks with maximum length = sequence max length of the H-network model.

Re-trained the h-network and the RL policy with the new reward function.

However, the policy still thinks keep charging all the time works the best. Maybe we have to standardize the input data to the policyas well.