Implement randomized episode for env
Instead of accpeting a dataframe containing a single episode (a day), the environment now accepts a custom dataloader object, which stores the whole train/test/validation dataframe and corresponding segments.

When an episode ends, the environment will call the dataloader to randomly load a segment, in a DataFrame format.

Also small code tidying by grouping the dataframe fragment and related datas (e.g. current timestep) to a SmartMeterEpisode object.