Fix f-signal implementation & H-network varient implementation
The f-signal should be a per-step expected value of log-likelihood, while previously we implemented f-signal as sum of all previous expected log-likelihood.

A varient of H-network is implemented. Instead of predicting the mean only with fixed 1 variance, we implemented a H-network which predicts both mean and variance.

This is because in the f-signal computation, the expectation will leads to the term \mathbb{E} [(x - \mu)^2] = \sigma^2, the variance, which leads to the entropy of a gaussian distribution only depends on the variance. Derivation is here: https://gregorygundersen.com/blog/2020/09/01/gaussian-entropy/

The following possible next steps are
- standardize the input to the NN as well
- implementing the cycle training process
- downsample the data to 1min data (instead of 1/6 Hz)