Unsuccessful result on training on standardized observation space
Attempt to train with standardized observation space to see if this can replicate the paper's result. Unfortunately this is also unsuccessful. The battery action remains simple after 150000 steps.

This commit serves as a record to the training result and modifications made.

Maybe we should move to integrating the training of LSTM H-network in the RL-environment.