Downsampling to 1min data
Write scripts to downsample the current 1yr dataset from 1/6 Hz to 1/60 Hz (i.e. 6 seconds per sample to 60 seconds per sample)

The 04_downsampling.ipynb is the script meant to be run after 03_data_split.ipynb

However, when executing the 04_downsampling.ipynb script, it is noticed that spliting the appliance's signature by train/test/validation dates in aggregate load DOES not work, as it breaks the appliance's signatures when a signature crosses the midnight.

There should be separate logic in handling splitting of appliance's load signature.