Implement episode identification & multi-day episodes training
With help of Copilot + Claude Code, we quickly implemented new data loaders which support single-day episodes, and multi-days episodes.

Both are equipped with better episode identification scheme, by computing a md5 hash based on episode starting datetime and ending datetime, and the length of the episode.

However, only the data loader from simple_data_loader.py is tested. It works perfectly.

The evaluation.ipynb, expt_results.ipynb, expt_results_validation.ipynb, and per_episode_analysis.ipynb are modified to support both data loader.

Refactoring is performed by categorizing the PPO training scheme, and the corresponding used h_network_rl_module, as the "continuous action space" training. This is because we are going to implement a baseline which is based on the paper "Privacy-Cost Management in Smart Meters With Mutual-Information-Based Reinforcement Learning". It uses a discretized state and action space, and there are significant difference in the h_network_rl_module.

Also, old files are removed for repo cleanliness.