Refactoring discrete & continuous action space scripts
Since we developed a baseline for discrete action space (i.e. our reproduction of DDQL-MI algorithm in the paper), most of the code are duplicate, especially in the environment itself.

Therefore, we opt for a refactoring on the RL environment to build a more unified training script/pipeline for our project, with the help of copilot + Claude

The battery, episode, h-network and the RL environment now follows a factory pattern.

We build a base class that defines shared functionalities for objects in discrete & continuous action space, and leave critical functions as abstract methods for further implementation.

Then for each object, we build one for continuous action space, and one for discrete action space, and implement those critical functions respectively.

To instantiate a correct object for the action space, the factory pattern can determine which type of object (continuous ver. or discrete ver.) should be created, allowing a unified function for both training scripts.

For revertability, this commit is the only commit that contains objects before refactoring and after refactoring. After development of a cleaner training script with simple env hyperparameter search, the old objects together with old training jupyter notebooks shall be removed.