RL env functionalities improvement
For discrete action-space environment, we add support to allow customizing step-size for user load / grid load quantization and battery charging/discharging quantization.

For both environments, we attempt to implement a battery difference signal for last step of an episode, to measure the difference between initial state-of-charge value, and final state-of-charge value, and we aim to minimize this. However, this value is very small, and finding the appropriate multiplier (like lambda in L2 regularization) will cost us lots of time (and we don't have time to do that). Hence, a blank function is left there.

Bug fix in discrete-action space env, for forgetting converting W to kW in calculating grid load.