Extend and Refactor on logging functionality, and more
We save more information after each episode in the environment for more detailed immediate analysis. This includes per-step overall reward, the f_signal and g_signal, and their sum, mean and std across the episode.

We move the actual file-saving logging call to a new callback. This is to support multiple environment training for PPO, which in thoery, should improve generalization performance of PPO.

This change involves creating new funcs in base class of the environment for performing log saving in the callback, storing last episode information temporarily in the env (for callback, as reset() is called before the callback, which those information may be overwritten before got saved), and logging .

Also, we added deterministicity in episode sampling in ValidateEveryNEpisodes for fair comparison between different models; and integration with Optuna for hyperparameter searching.