Refactoring on the training scripts.
This is to adopt the refactored changes in the env classes and h-network classes.

We first adopt the changes to the ipynb, then modified the .py training scripts. The .py scripts are superior to the .ipynb notebooks, as they are tested for longer training, and they are expected to be executed via a bash script.

Also, we introduced seed for reproducibility. In short, the seed is to control the env episode sampling, action space sampling, and network (RL agent and h-network) weight initialization. For the last item, reproducibility is not guaranteed due to architecture/software version differences. However, a consistent env episode sampling should lead to more fair comparisons between different training runs (of different RL agents).