Optuna hyperparameter searching
To optimize our algorithm, we make use of Optuna for hyperparameter searching.

We search on lr, batch_size, n_steps, n_epochs, gae_lambda, activation_func (of the underhood NN) of the agent; and search on lr of the H-network. This is to find the optimal hyperparameter for our network.

Note that to ensure fair comparison between different hyparameters, a constant seed is being set to ensure same episode sampling (of training env and validation env), same action_space sampling, and hopefully, same initialization on NN of RL agent and H-network. While this may include bias, this is for fair comparison between different algorithms (PPO vs DDQL-MI)