Fix incorrect plot for H-network loss analysis under discrete action space
Previously we mistakenly use the same GaussianNLLLoss for the discrete case, which does not reflect the discrete action space assumption. We fix this with an approximated cross-entropy loss based on the argmax class and the true household load class (as we do not store the whole probability due to memory issue)

We also fix the incorrect y-axis in ax1 for the same plot under the discrete action space case.